Squashed commit of the following from refact/inoutstream:
- Fix bug with lock and unlock methods
- Change the key parameter type from uint8_t to uint32_t
- Remove the callback methods meant to be used with inheritance
- Add the definition for a invalid dio
- Fix some guard clauses
- Create a separate file for inoutstream data types
- Fix typos
- Change inoutstream's file and class name

+++ drivers/drivers/linux/utils/linux_types.hpp
@@ -35,7 +35,7 @@ typedef struct
   bool terminate;
   uint8_t *buffer;
   uint32_t size;
-  uint8_t key;
+  uint32_t key;
   InOutStreamCallback_t func;
   void *arg;
 }UtilsInOutSync_t;
